zcash_client_sqlite: Move testing helper functions into module
